Remove separate export page
